Graph plugin: convert to Slang-generated shaders, packed-u32 color
graph.slang is the source of truth for WGSL + GLSL and TS layout
constants (stride, field offsets, attributes). Color moves to a single
packed ABGR u32 ATTR, unpacked in the shader; CPU-side color generation
produces u32 directly so there's no float→u32 repacking at upload time.

SubBatch is now a single interleaved vertex buffer (f32 + u32 view over
the same ArrayBuffer) driven by FIELD_OFFSET_F32 / INSTANCE_STRIDE_*.
MeshBuilder grows a typed buffer in place so toSubBatch is one slice
instead of a full interleave pass. Renderers consolidate
edge/node/arrow fields into a Record<SubBatchKey, …> to kill three
copies of the same ternary. Dead recolorNodes removed.

+++ plugins/graph/src/GraphGenomeView/model.ts
@@ -47,9 +47,9 @@ export default function stateModelFactory() {
       nodeVertexRanges: undefined as Map<string, VertexRange> | undefined,
       edgeVertexRanges: undefined as Map<number, VertexRange> | undefined,
       arrowVertexRanges: undefined as Map<number, VertexRange> | undefined,
-      baseNodeColors: undefined as Float32Array | undefined,
-      baseEdgeColors: undefined as Float32Array | undefined,
-      baseArrowColors: undefined as Float32Array | undefined,
+      baseNodeColors: undefined as Uint32Array | undefined,
+      baseEdgeColors: undefined as Uint32Array | undefined,
+      baseArrowColors: undefined as Uint32Array | undefined,
     }))
     .views(self => ({
       get nodeCount() {
@@ -119,9 +119,9 @@ export default function stateModelFactory() {
         nodeVertexRanges: Map<string, VertexRange>,
         edgeVertexRanges: Map<number, VertexRange>,
         arrowVertexRanges: Map<number, VertexRange>,
-        baseNodeColors: Float32Array,
-        baseEdgeColors: Float32Array,
-        baseArrowColors: Float32Array,
+        baseNodeColors: Uint32Array,
+        baseEdgeColors: Uint32Array,
+        baseArrowColors: Uint32Array,
       ) {
         self.nodeVertexRanges = nodeVertexRanges
         self.edgeVertexRanges = edgeVertexRanges
